Move logging back under utilities for now, minimizing the change
