+++ src/tiledb/cloud/files/indexing.py
@@ -3,8 +3,8 @@ from typing import Any, Dict, List, Mapping, Optional, Sequence
 
 import tiledb
 from tiledb.cloud import dag
-from tiledb.cloud.logging import get_logger_wrapper
 from tiledb.cloud.utilities import as_batch
+from tiledb.cloud.utilities.logging import get_logger_wrapper
 
 
 class IndexTypes(str, Enum):
