+++ src/tiledb/cloud/utilities/__init__.py
@@ -1,3 +1,5 @@
+"""Utility functions."""
+
 from ._common import as_batch
 from ._common import chunk
 from ._common import find
@@ -11,6 +13,8 @@ from ._common import set_aws_context
 from .consolidate import consolidate_and_vacuum
 from .consolidate import consolidate_fragments
 from .consolidate import group_fragments
+from .logging import get_logger
+from .logging import get_logger_wrapper
 from .profiler import Profiler
 from .profiler import create_log_array
 from .profiler import write_log_event
@@ -21,6 +25,8 @@ __all__ = [
     "as_batch",
     "chunk",
     "find",
+    "get_logger",
+    "get_logger_wrapper",
     "max_memory_usage",
     "process_stream",
     "read_aws_config",
