+++ src/tiledb/cloud/utilities/wheel.py
@@ -14,7 +14,7 @@ from typing import Any, List, Mapping, Optional, Sequence
 
 import tiledb
 from tiledb.cloud.client import Config
-from tiledb.cloud.logging import get_logger
+from tiledb.cloud.utilities.logging import get_logger
 
 logger = get_logger()
 logger.propagate = False
